+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added `BuildContext.DeferPostprocessAsset` (#229)
+
+### Changed
+
+### Deprecated
+
+### Removed
+- `BuildContext.Serialize()`, which was accidentally made public for use in a unit test, is now an `[Obsolete]` no-op.
+
+### Fixed
+
+### Security
+
+
 ## [1.4.0] - [2024-03-27]
 
 ### Added